+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Braintree Android Drop-In Release Notes
 
-## unreleased
+## 6.16.0
 
 * Bump braintree_android module dependency versions to `4.45.0`
 * Fixes Google Play Store Rejection